Use separate module for base class
